CP-90 sbc_compassion installable
