+++ sbc_compassion/views/download_letters_view.xml
@@ -23,9 +23,8 @@
     <record id="action_download_letter" model="ir.actions.act_window">
         <field name="name">Download archive</field>
         <field name="res_model">correspondence.download.wizard</field>
-        <field name="binding_model">correspondence</field>
+        <field name="binding_model_id" ref="model_correspondence"/>
         <field name="view_mode">form</field>
         <field name="target">new</field>
-        <field name="key2">client_action_multi</field>
     </record>
 </odoo>
