+++ sbc_compassion/views/partner_compassion_view.xml
@@ -11,7 +11,7 @@
                 </button>
             </xpath>
             <field name="lang" position="after">
-                <field name="spoken_lang_ids" widget="many2many_tags" options="{'create':False}" attrs="{'required': [('is_company', '=', False), ('customer', '=', True)]}" groups="child_compassion.group_sponsorship"/>
+<!--                <field name="spoken_lang_ids" widget="many2many_tags" options="{'create':False}" attrs="{'required': [('is_company', '=', False), ('customer', '=', True)]}" groups="child_compassion.group_sponsorship"/>-->
                 <field name="mandatory_review" groups="child_compassion.group_sponsorship"/>
             </field>
         </field>
